+++ CatsAndDogs/src/main/java/com/hanghae99/catsanddogs/service/PostService.java
@@ -1,11 +1,8 @@
 package com.hanghae99.catsanddogs.service;
 
-import com.hanghae99.catsanddogs.dto.LikeCommentResponseDto;
-import com.hanghae99.catsanddogs.dto.LikePostResponseDto;
 import com.hanghae99.catsanddogs.dto.comment.CommentResponseDto;
 import com.hanghae99.catsanddogs.dto.post.*;
 import com.hanghae99.catsanddogs.entity.Comment;
-import com.hanghae99.catsanddogs.entity.LikeComment;
 import com.hanghae99.catsanddogs.entity.Post;
 import com.hanghae99.catsanddogs.entity.User;
 import com.hanghae99.catsanddogs.exception.CustomException;
@@ -14,18 +11,14 @@ import com.hanghae99.catsanddogs.repository.CommentRepository;
 import com.hanghae99.catsanddogs.repository.LikeCommentRepository;
 import com.hanghae99.catsanddogs.repository.LikePostRepository;
 import com.hanghae99.catsanddogs.repository.PostRepository;
-import com.hanghae99.catsanddogs.s3.S3Uploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
 @Service
@@ -40,23 +33,16 @@ public class PostService {
     private final LikeCommentRepository likeCommentRepository;
 
 
-    @Autowired
-    private S3Uploader s3Uploader;
-
-
     @Transactional(readOnly = true)
     public MainResponseDto getPostList(Long userId) {
 
-        List<Post> postList = postRepository.findAll(); //게시물이 많아지면 문제가 될듯
+        List<Post> postList = postRepository.findAll();
         PostResponseListDto postResponseListDto = new PostResponseListDto();
         for(Post post : postList){
             postResponseListDto.addPostResponseDto(new PostResponseDto(post, userId));
         }
         List<PostResponseDto> rank = postRepository.findTop2ByCreatedAtAfterOrderByLikeCountDesc(LocalDateTime.now().minusDays(1)).stream().map(post -> new PostResponseDto(post, userId)).collect(Collectors.toList());
-        MainResponseDto mainResponseDto = new MainResponseDto(postResponseListDto, rank);
-
-        LocalDateTime.now().minusDays(1);
-        return mainResponseDto;
+        return new MainResponseDto(postResponseListDto, rank);
     }
 
     @Transactional(readOnly = true)
@@ -71,18 +57,10 @@ public class PostService {
 
 
     @Transactional
-    public PostResponseDto createPost(PostRequestDto requestDto, MultipartFile image, User user) throws Exception {
-        if(image.getContentType() == null || !image.getContentType().startsWith("image"))
-            throw new CustomException(ErrorCode.WRONG_IMAGE_FORMAT);
+    public PostResponseDto createPost(PostRequestDto requestDto, User user) {
 
-        UUID uuid = UUID.randomUUID(); // 파일 이름에 붙일 랜덤 식별자
-        String pictureName = uuid + "_" + image.getOriginalFilename(); // 새로운 이름 - 이름이 같으면 오류나서 이렇게 해줌
 
-        if(!image.isEmpty()) {
-            String storedFileName = s3Uploader.upload(image,"images");
-            requestDto.setPicturePath(storedFileName);
-        }
-        Post post = postRepository.save(new Post(requestDto, user, requestDto.getPicturePath(), "pictureName"));
+        Post post = postRepository.save(new Post(requestDto, user));
 
         return new PostResponseDto(post);
 
@@ -91,38 +69,25 @@ public class PostService {
 
 
     @Transactional
-    public PostResponseDto updatePost(Long postId, PostRequestDto requestDto, MultipartFile image, User user) throws Exception {
+    public PostResponseDto updatePost(Long postId, PostRequestDto requestDto, User user)  {
 
-        if(!image.getContentType().startsWith("image"))
-            throw new CustomException(ErrorCode.WRONG_IMAGE_FORMAT);
-
-        UUID uuid = UUID.randomUUID(); // 파일 이름에 붙일 랜덤 식별자
-        String pictureName = uuid + "_" + image.getOriginalFilename(); // 새로운 이름 - 이름이 같으면 오류나서 이렇게 해줌
-
-        if(!image.isEmpty()) {
-            String storedFileName = s3Uploader.upload(image,"images");
-            requestDto.setPicturePath(storedFileName);
-        }
-
-//        Post post = postRepository.findByIdAndUsers(postId, user).orElseThrow(
-//                () -> new CustomException(ErrorCode.CONTENT_NOT_FOUND)
-//        );
 
         Post post = postRepository.findById(postId).orElseThrow(()-> new CustomException(ErrorCode.CONTENT_NOT_FOUND));
 
         if(!post.getUsers().getId().equals(user.getId())){
             throw new CustomException(ErrorCode.AUTHORIZATION_UPDATE_FAIL);
         }
-        post.update(requestDto, requestDto.picturePath, pictureName);
+        post.update(requestDto);
 
         return new PostResponseDto(post);
     }
 
 
     @Transactional
-    public PostResponseDto deletePost(Long postId, User user) {
+    public void deletePost(Long postId, User user) {
 
         Post post = postRepository.findById(postId).orElseThrow(()-> new CustomException(ErrorCode.CONTENT_NOT_FOUND));
+
         if(!post.getUsers().getId().equals(user.getId())) {
             throw new CustomException(ErrorCode.AUTHORIZATION_DELETE_FAIL);
         }
@@ -132,11 +97,14 @@ public class PostService {
             commentIdList.add(comment.getId());
         }
 
-        likeCommentRepository.deleteAllByCommentIdIn(commentIdList);
-        commentRepository.deleteAllByIdIn(commentIdList);
+        if(!commentIdList.isEmpty()){
+            likeCommentRepository.deleteAllByCommentIdIn(commentIdList);
+            commentRepository.deleteAllByIdIn(commentIdList);
+        }
+
         likePostRepository.deleteAllByPostId(postId);
         postRepository.delete(post);
-        return new PostResponseDto();
+
 
     }
 }
